+++ lib/data/enum/admin_access.dart
@@ -6,4 +6,5 @@ enum AdminAccess {
   vacation,
   cra,
   wordpressUsers,
+  featureFlipping,
 }