+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 _Chore_
 
 - Use the Github worflows from Xpeho Workflows repository
+- [Issue #33](https://github.com/XPEHO/xpeapp_admin/issues/33) Update icon and name of the app
 
 ## 0.4.0